Регистрация и аутентификация

+++ homework_3/src/main/java/ru/parfenov/homework_3/servlets/admin/DeleteUserServlet.java
@@ -4,7 +4,10 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
+import ru.parfenov.homework_3.enums.UserRole;
+import ru.parfenov.homework_3.model.User;
 import ru.parfenov.homework_3.service.UserService;
 import ru.parfenov.homework_3.utility.Utility;
 
@@ -21,9 +24,19 @@ public class DeleteUserServlet extends HttpServlet {
 
     @Override
     public void doDelete(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        String userIdStr = request.getParameter("id");
-        String jsonString = userService.delete(userIdStr) ? "the user is deleted" : "the user is not deleted!";
-        response.setStatus("the user is not deleted!".equals(jsonString) ? 404 : 200);
+        HttpSession session = request.getSession();
+        int responseStatus;
+        var user = (User) session.getAttribute("user");
+        String jsonString;
+        if (user == null || user.getRole() != UserRole.ADMIN) {
+            jsonString = "no rights or registration!";
+            responseStatus = user == null ? 401 : 403;
+        } else {
+            String userIdStr = request.getParameter("id");
+            jsonString = userService.delete(userIdStr) ? "the user is deleted" : "the user is not deleted!";
+            responseStatus = "the user is not deleted!".equals(jsonString) ? 404 : 200;
+        }
+        response.setStatus(responseStatus);
         PrintWriter out = response.getWriter();
         response.setContentType("application/json");
         response.setCharacterEncoding("UTF-8");
