+++ homework_3/src/main/java/ru/parfenov/homework_3/servlets/client/DeleteCarServlet.java
@@ -4,7 +4,9 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
+import ru.parfenov.homework_3.model.User;
 import ru.parfenov.homework_3.service.CarService;
 import ru.parfenov.homework_3.utility.Utility;
 
@@ -21,9 +23,19 @@ public class DeleteCarServlet extends HttpServlet {
 
     @Override
     public void doDelete(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        String carIdStr = request.getParameter("id");
-        String jsonString = carService.delete(carIdStr) ? "the car is deleted" : "the car is not deleted!";
-        response.setStatus("the car is not deleted!".equals(jsonString) ? 404 : 200);
+        HttpSession session = request.getSession();
+        int responseStatus;
+        var user = (User) session.getAttribute("user");
+        String jsonString;
+        if (user == null) {
+            jsonString = "no registration!";
+            responseStatus = 401;
+        } else {
+            String carIdStr = request.getParameter("id");
+            jsonString = carService.delete(carIdStr) ? "the car is deleted" : "the car is not deleted!";
+            responseStatus = "the car is not deleted!".equals(jsonString) ? 404 : 200;
+        }
+        response.setStatus(responseStatus);
         PrintWriter out = response.getWriter();
         response.setContentType("application/json");
         response.setCharacterEncoding("UTF-8");
